# api-gateway/api-gateway-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/job/create": {
            "post": {
                "description": "Creates and schedule a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Creates and schedule a job",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateJobResponse"
                        }
                    }
                }
            }
        },
        "/job/delete": {
            "post": {
                "description": "Delete a job by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Delete a job by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "some id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.DeleteJobResponse"
                        }
                    }
                }
            }
        },
        "/job/get": {
            "get": {
                "description": "Get a job by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Get a job by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "some id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.Job"
                        }
                    }
                }
            }
        },
        "/job/list": {
            "get": {
                "description": "list jobs by type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "list jobs by type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListJobsResponse"
                        }
                    }
                }
            }
        },
        "/job/update": {
            "post": {
                "description": "Updates and reschedule a job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "job"
                ],
                "summary": "Updates and reschedule a job",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.UpdateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.UpdateJobResponse"
                        }
                    }
                }
            }
        },
...
        "/report/list": {
            "get": {
                "description": "retrieve the reports",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "report"
                ],
                "summary": "Get the list of reports",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListReportResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateJobRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "jobData": {
                    "type": "string"
                },
                "jobType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scheduleTime": {
                    "type": "string"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.CreateJobResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.DeleteJobResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.Job": {
            "type": "object",
            "properties": {
                "createdTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobData": {
                    "type": "string"
                },
                "jobStatus": {
                    "type": "integer"
                },
                "jobType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scheduleTime": {
                    "type": "string"
                },
                "updatedTime": {
                    "type": "string"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListJobsResponse": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.Job"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.ListReportResponse": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.Report"
                    }
                },
                "size": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.Report": {
            "type": "object",
            "properties": {
                "createdTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "reportData": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.UpdateJobRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobData": {
                    "type": "string"
                },
                "jobType": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "scheduleTime": {
                    "type": "string"
                }
            }
        },
        "github_com_mikemonzo_play-microservices_api-gateway_api-gateway-service_internal_models.UpdateJobResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}